requirements.txt
PyPDF2
discord.py
pytz
pyxdg
